Add if isset

+++ app/task/delete.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 require __DIR__ . '/../autoload.php';
 
-$id = $_POST['delete-task'];
+if (isset($_POST['delete-task'])) {
+    $id = $_POST['delete-task'];
 
-$statement = $database->prepare('DELETE FROM tasks WHERE id = :id;');
+    $statement = $database->prepare('DELETE FROM tasks WHERE id = :id;');
 
-$statement->bindParam(':id', $id, PDO::PARAM_INT);
+    $statement->bindParam(':id', $id, PDO::PARAM_INT);
 
-$statement->execute();
-
-
+    $statement->execute();
+};
 redirect('/list.php');
